Bring back calendar page, convert page into loading.tsx, finishes loading for Calendar view
