--- src/app/home/calendar/real/page.tsx
@@ -1,11 +1,0 @@
-import { Header } from '@/components/Header';
-import { CalendarView } from '../calendar';
-
-export default function Page() {
-  return (
-    <>
-      <Header text='Calendar' />
-      <CalendarView />
-    </>
-  )
-}
